Students can see,delete details for just theirselves.

+++ src/main/java/com/example/UniversityInformationSystem/controller/StudentController.java
@@ -2,7 +2,9 @@ package com.example.UniversityInformationSystem.controller;
 
 import com.example.UniversityInformationSystem.dto.response.CourseDto;
 import com.example.UniversityInformationSystem.dto.response.StudentDto;
+import com.example.UniversityInformationSystem.exception.LogicalMistakeException;
 import com.example.UniversityInformationSystem.model.CourseModel;
+import com.example.UniversityInformationSystem.security.JwtFilter;
 import com.example.UniversityInformationSystem.service.CourseService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -26,6 +28,7 @@ public class StudentController {
 
     private List<CourseDto> courseDtoList;
 
+    private final JwtFilter jwtFilter;
 
     @GetMapping
     public ResponseEntity<String> merhaba(){
@@ -42,12 +45,18 @@ public class StudentController {
     @GetMapping("/{studentId}")
     public ResponseEntity<StudentDto> getStudentById(@PathVariable Long studentId){
 
+        studentValidator(studentId);
+
         return new ResponseEntity<StudentDto>(studentService.convertToStudentDto(studentService.getStudentById(studentId)), HttpStatus.OK);
 
     }
 
     @GetMapping("/courses/{studentId}")
     public ResponseEntity<List<CourseDto>> getTakenCourses(@PathVariable Long studentId){
+
+        studentValidator(studentId);
+
+
         courseModelList.clear();
         courseDtoList.clear();
         courseModelList = studentService.getTakenCourses(studentId);
@@ -65,6 +74,12 @@ public class StudentController {
     @DeleteMapping("/delete/{studentId}")
     public ResponseEntity<String> deleteStudentById(@PathVariable Long studentId){
 
+        String role = jwtFilter.getUser().getAuthorities().toString();
+
+        if(role.equals("[Student]")){
+            throw new LogicalMistakeException("You can not delete your account! Please contact to admin.");
+        }
+
         studentService.deleteStudentById(studentId);
         return new ResponseEntity<String>("Student deleted successfully!",HttpStatus.OK);
 
@@ -73,6 +88,10 @@ public class StudentController {
     @DeleteMapping("/delete/course/{studentId}-{courseId}")
     public ResponseEntity<List<CourseDto>> deleteCourseFromTakens(@PathVariable Long studentId
                                                                 , @PathVariable Long courseId){
+
+
+        studentValidator(studentId);
+
         courseModelList.clear();
         courseDtoList.clear();
         courseModelList = studentService.deleteCourseFromTakens(studentId,courseId);
@@ -82,6 +101,21 @@ public class StudentController {
         });
 
         return new ResponseEntity<List<CourseDto>>(courseDtoList,HttpStatus.OK);
+    }
+
+
+    public boolean studentValidator(Long studentId){
+        System.out.println(jwtFilter.getUser().getUsername());
+        String role = jwtFilter.getUser().getAuthorities().toString();
+
+        if(role.equals("[Student]")){
+            if (!studentService.getStudentById(studentId).getEmail().equals(jwtFilter.getUser().getUsername())){
+
+                throw new LogicalMistakeException("You can look/edit details only for yourself!");
+
+            }
+        }
+        return true;
     }
